Add new markdown file for 2404.16801 paper announcement
This commit introduces a new markdown file titled
'2024-04-25-2404.16801.md', which serves as an announcement for the
publication of the paper "The Directed Landscape is a Black Noise" by
Zoe Himwich and Shalin Parekh on arXiv. The file includes standard front
matter such as layout, title, publication status, and tags, along with
links to the paper on arXiv.